backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.0.3
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
websockets==12.0
python-multipart==0.0.6
apscheduler==3.10.4
anthropic==0.7.8
openai==1.3.7
boto3==1.34.0
botocore==1.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
aiofiles==23.2.1
aiohttp==3.9.1
Brotli==1.1.0 